+++ cmd/auth/auth.go
@@ -16,29 +16,31 @@
 	    along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 
-package cmd
+package auth
 
 import (
+	"github.com/matthieukhl/rgvr/cmd"
 	"github.com/spf13/cobra"
 )
 
-// authCmd represents the auth command
-var authCmd = &cobra.Command{
+// AuthCmd represents the auth command
+var AuthCmd = &cobra.Command{
 	Use:   "auth",
 	Short: "Login, logout, and manage your Ringover API key",
 	Long:  `The auth command allows you to log in, log out, and manage your Ringover API key.`,
 }
 
-func init() {
-	rootCmd.AddCommand(authCmd)
 
+func init() {
+	cmd.RootCmd.AddCommand(AuthCmd)
+	
 	// Here you will define your flags and configuration settings.
 
 	// Cobra supports Persistent Flags which will work for this command
 	// and all subcommands, e.g.:
-	authCmd.PersistentFlags().String("login", "", "Register your Ringover API key")
+	AuthCmd.PersistentFlags().String("login", "", "Register your Ringover API key")
 
 	// Cobra supports local flags which will only run when this command
 	// is called directly, e.g.:
-	// authCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
+	// AuthCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
 }