+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.3.0-beta] - 2026-07-31
 
 ### Added
 
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `users get <user_id>` command to retrieve a user's information based on a user ID.
 - Added `users plannings get <user_id>` command to retrieve a user's planning based on a user ID.
 - Added `users presences <user_id>` command to retrieve a user's presence status based on a user ID.
+- Added `users snooze log <user_id>` command to retrieve a specific user's snooze log.
 
 ### Modified